# main.go
package main

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"
	"time"

	"nwd-deakr/internal/adapters/benefitsindex"
	"nwd-deakr/internal/adapters/residentindex"
	"nwd-deakr/internal/domain"
)

const unifiedRequestTimeout = 10 * time.Second

var (
	residentIndexURL = "http://127.0.0.1:8081"
	benefitsURL      = "http://127.0.0.1:8082"
)

var client = &http.Client{
	Timeout: 5 * time.Second,
}

func getResident(client *residentindex.Client) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		id := r.PathValue("id")

		if id == "" {
			http.Error(w, "resident ID is required", http.StatusBadRequest)
			return
		}

		resident, status := client.GetResident(r.Context(), id)

		if status.Status == "not_found" {
			http.Error(w, "resident not found", http.StatusNotFound)
			return
		}

		if status.Status != "ok" {
			http.Error(w, "resident index unavailable", http.StatusBadGateway)
			return
		}

		w.Header().Set("Content-Type", "application/json")
		json.NewEncoder(w).Encode(resident)
	}
}

func getBenefit(client *benefitsindex.Client) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		ref := r.PathValue("ref")

		if ref == "" {
			http.Error(w, "benefit reference is required", http.StatusBadRequest)
			return
		}

		record, status := client.GetBenefit(r.Context(), ref)

		if status.Status == "not_found" {
			http.Error(w, "benefit record not found", http.StatusNotFound)
			return
		}

		if status.Status != "ok" {
			http.Error(w, "benefits register unavailable", http.StatusBadGateway)
			return
		}

		w.Header().Set("Content-Type", "application/json")
		json.NewEncoder(w).Encode(record)
	}
}

func getResidents(client *residentindex.Client) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		residents, pagination, status := client.GetResidents(r.Context())

		response := domain.ResidentListResponse{
			Residents:  residents,
			Pagination: pagination,
			Meta: domain.UnifiedMeta{
				Sources: map[string]domain.SourceStatus{
					"resident_index": status,
				},
				Partial: status.Status != "ok" || !pagination.Complete,
			},
		}

		w.Header().Set("Content-Type", "application/json")
		json.NewEncoder(w).Encode(response)
	}
}

func getUnified(
	residentClient *residentindex.Client,
	benefitsClient *benefitsindex.Client,
) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		residentID := r.URL.Query().Get("resident_id")
		benefitRef := r.URL.Query().Get("benefit_ref")

		if residentID == "" && benefitRef == "" {
			http.Error(
				w,
				"resident_id or benefit_ref is required",
				http.StatusBadRequest,
			)
			return
		}

		ctx, cancel := context.WithTimeout(r.Context(), unifiedRequestTimeout)
		defer cancel()

		type residentResult struct {
			data   *domain.Resident
			status domain.SourceStatus
		}

		type benefitResult struct {
			data   *domain.BenefitRecord
			status domain.SourceStatus
		}

		residentCh := make(chan residentResult, 1)
		benefitCh := make(chan benefitResult, 1)

		if residentID != "" {
			go func() {
				data, status := residentClient.GetResident(
					ctx,
					residentID,
				)

				residentCh <- residentResult{
					data:   data,
					status: status,
				}
			}()
		}

		if benefitRef != "" {
			go func() {
				data, status := benefitsClient.GetBenefit(
					ctx,
					benefitRef,
				)

				benefitCh <- benefitResult{
					data:   data,
					status: status,
				}
			}()
		}

		response := domain.UnifiedResponse{
			Meta: domain.UnifiedMeta{
				Sources: make(map[string]domain.SourceStatus),
			},
		}

		if residentID != "" {
			result := <-residentCh

			response.Resident = result.data
			response.Meta.Sources["resident_index"] = result.status

			if result.status.Status != "ok" {
				response.Meta.Partial = true
			}
		}

		if benefitRef != "" {
			result := <-benefitCh

			response.Benefits = result.data
			response.Meta.Sources["benefits_register"] = result.status

			if result.status.Status != "ok" {
				response.Meta.Partial = true
			}
		}

		w.Header().Set("Content-Type", "application/json")

		json.NewEncoder(w).Encode(response)
	}
}

func getResidentUnified(
	residentClient *residentindex.Client,
	catalogue *benefitsindex.Catalogue,
) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		id := r.PathValue("id")

		if id == "" {
			http.Error(w, "resident ID is required", http.StatusBadRequest)
			return
		}

		resident, resStatus := residentClient.GetResident(r.Context(), id)

		if resStatus.Status == "not_found" {
			http.Error(w, "resident not found", http.StatusNotFound)
			return
		}

		response := domain.AutoUnifiedResponse{
			Resident: resident,
			IdentityMatch: domain.IdentityMatchMeta{
				Outcome:       domain.IdentityUnavailable,
				CandidateRefs: []string{},
				Evidence:      []domain.IdentityEvidence{},
			},
			Meta: domain.UnifiedMeta{
				Sources: map[string]domain.SourceStatus{
					"resident_index": resStatus,
				},
			},
		}

		records, fresh, fetchedAtMs := catalogue.Get(r.Context())

		if records == nil {
			response.Meta.Sources["benefits_register"] = domain.SourceStatus{
				Status:       "unavailable",
				ErrorMessage: "benefits catalogue unavailable",
			}
		} else {
			benefitSource := domain.SourceStatus{Status: "ok"}
			if !fresh {
				benefitSource.Status = "stale"
				benefitSource.ErrorMessage = "serving cached catalogue snapshot"
			}
			response.Meta.Sources["benefits_register"] = benefitSource

			matched, matchMeta := domain.MatchResidentToCatalogue(resident, records, fetchedAtMs)
			response.Benefits = matched
			response.IdentityMatch = matchMeta
		}

		benefitsStatus := response.Meta.Sources["benefits_register"]
		response.Meta.Partial = resStatus.Status != "ok" || benefitsStatus.Status != "ok"

		w.Header().Set("Content-Type", "application/json")
		json.NewEncoder(w).Encode(response)
	}
}

func getHealth(
	residentClient *residentindex.Client,
	benefitsClient *benefitsindex.Client,
) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		ctx, cancel := context.WithTimeout(r.Context(), 3*time.Second)
		defer cancel()

		type healthResult struct {
			name   string
			status domain.SourceStatus
		}
		ch := make(chan healthResult, 2)

		go func() { ch <- healthResult{"resident_index", residentClient.GetHealth(ctx)} }()
		go func() { ch <- healthResult{"benefits_register", benefitsClient.GetHealth(ctx)} }()

		sources := make(map[string]domain.SourceStatus, 2)
		for range 2 {
			res := <-ch
			sources[res.name] = res.status
		}

		overall := "ok"
		for _, s := range sources {
			if s.Status != "ok" {
				overall = "degraded"
				break
			}
		}

		w.Header().Set("Content-Type", "application/json")
		json.NewEncoder(w).Encode(domain.HealthResponse{
			Status:  overall,
			Sources: sources,
		})
	}
}

func main() {
	residentClient := residentindex.NewClient(
		residentIndexURL,
		client,
	)

	benefitsClient := benefitsindex.NewClient(
		benefitsURL,
		client,
	)

	catalogue := benefitsindex.NewCatalogue(benefitsClient, 0)

	mux := http.NewServeMux()

	mux.HandleFunc(
		"GET /residents/{id}",
		getResident(residentClient),
	)

	mux.HandleFunc(
		"GET /residents/{id}/unified",
		getResidentUnified(residentClient, catalogue),
	)

	mux.HandleFunc(
		"GET /residents",
		getResidents(residentClient),
	)

	mux.HandleFunc(
		"GET /benefits/{ref...}",
		getBenefit(benefitsClient),
	)

	mux.HandleFunc(
		"GET /docs",
		serveDocs,
	)

	mux.HandleFunc(
		"GET /docs/swagger-ui.css",
		serveSwaggerCSS,
	)

	mux.HandleFunc(
		"GET /docs/swagger-ui-bundle.js",
		serveSwaggerJS,
	)

	mux.HandleFunc(
		"GET /openapi.yaml",
		serveOpenAPISpec,
	)

	mux.HandleFunc(
		"GET /unified",
		getUnified(residentClient, benefitsClient),
	)

	mux.HandleFunc(
		"GET /health",
		getHealth(residentClient, benefitsClient),
	)

	fmt.Println("Unified API running on http://127.0.0.1:8080")

	if err := http.ListenAndServe(":8080", mux); err != nil {
		fmt.Println("Server stopped:", err)
	}
}