# openapi.yaml
openapi: 3.0.3
info:
  title: No Wrong Door — Unified Resident Integration API
  version: 1.0.0
  description: |
    One integration API that unifies citizen data across two legacy municipal sources:
    a paginated REST Resident Index and an unreliable XML Benefits Register.

    ### Key Architectural Behaviors
    * **Graceful Degradation:** Aggregate endpoints always respond with HTTP 200 and a structured `_meta` diagnostic envelope explaining any failed upstream sources rather than failing the caller.
    * **Catalogue Deduplication:** The Resident Index pagination shifts boundaries dynamically; duplicates are removed by stable resident ID (preserving first-seen order) and verified against the source total.
    * **Failure Resilience:** The Benefits Register operates under a 40% upstream failure rate; calls are protected with bounded retries, an in-memory TTL cache, and a circuit breaker.

servers:
  - url: http://127.0.0.1:8080
    description: Local Unified API Server

tags:
  - name: Unified Integration
    description: Primary aggregation endpoint returning a combined citizen view with source diagnostics.
  - name: Resident Index
    description: REST service (Port 8081) for individual resident lookups and full catalogue deduplication.
  - name: Benefits Register
    description: XML service (Port 8082) for legacy entitlement and benefit claim records.
  - name: Operations
    description: Service health and observability.

paths:
  /health:
    get:
      tags:
        - Operations
      summary: Aggregate upstream health check
      description: |
        Probes both upstream services concurrently (3 s timeout) and returns their live status.
        For the Benefits Register, also reports the current circuit breaker state (`circuit`).
        Overall `status` is `ok` only when both sources respond healthy; otherwise `degraded`.
      operationId: getHealth
      responses:
        '200':
          description: Health status of all upstream sources.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
              example:
                status: ok
                sources:
                  resident_index:
                    status: ok
                    http_code: 200
                    latency_ms: 3
                  benefits_register:
                    status: ok
                    circuit: closed
                    http_code: 200
                    latency_ms: 6


  /unified:
    get:
      tags:
        - Unified Integration
      summary: Unified view across both sources
      description: |
        Queries either or both upstream sources concurrently and returns everything known about
        the person, including per-source status metadata. Provide at least one query parameter.

        This endpoint never fails because an upstream failed: failures are reported inside
        `_meta.sources` with `partial: true` while any healthy source data is returned.
      operationId: getUnifiedView
      parameters:
        - name: resident_id
          in: query
          required: false
          description: Resident Index identifier to look up.
          schema:
            type: string
          example: R-10001
        - name: benefit_ref
          in: query
          required: false
          description: Benefits Register reference to look up (multi-segment path).
          schema:
            type: string
          example: CA/2016/4001
      responses:
        '200':
          description: Unified record with per-source status diagnostics.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnifiedResponse'
        '400':
          description: Neither resident_id nor benefit_ref was provided.

  /residents/{id}/unified:
    get:
      tags:
        - Unified Integration
      summary: Automatic cross-source identity resolution
      description: |
        Looks up one resident, then searches the cached Benefits Register
        catalogue for that same person using an evidence-based matcher.
        The sources share no key, so linking is performed on normalized
        identity attributes (exact date of birth plus name; falling back to
        name plus town plus street address).

        The answer always carries its evidence. A single rule match merges;
        multiple identical candidates are reported as ambiguous and never
        merged; zero candidates is a no_match; if the catalogue cannot be
        obtained at all the outcome is unavailable — never silently no_match.

        Ground truth (measured against the data pack's hidden person IDs):
        precision 100%, recall 100% across all 340 known pairs.
      operationId: getResidentAutoUnified
      parameters:
        - name: id
          in: path
          required: true
          description: Resident Index identifier.
          schema:
            type: string
          example: R-10001
      responses:
        '200':
          description: Resident view with matched benefit record and identity receipt.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutoUnifiedResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  /residents:
    get:
      tags:
        - Resident Index
      summary: Fetch full deduplicated resident catalogue
      description: |
        Walks every page of the Resident Index sequentially, removes duplicates
        caused by shifting page boundaries, and returns the full catalogue together
        with a pagination receipt describing collection statistics.
      operationId: listResidents
      responses:
        '200':
          description: Deduplicated catalogue with pagination receipt and source metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResidentListResponse'

  /residents/{id}:
    get:
      tags:
        - Resident Index
      summary: Fetch a single resident
      description: Looks up one record in the Resident Index by its stable ID.
      operationId: getResidentById
      parameters:
        - name: id
          in: path
          required: true
          description: Resident Index identifier.
          schema:
            type: string
          example: R-10001
      responses:
        '200':
          description: The resident record.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resident'
        '404':
          $ref: '#/components/responses/NotFound'
        '502':
          $ref: '#/components/responses/UpstreamUnavailable'

  /benefits/{ref}:
    get:
      tags:
        - Benefits Register
      summary: Fetch a single benefits record
      description: |
        Looks up one record in the legacy XML Benefits Register by reference.
        References contain slashes (e.g. CA/2016/4001). Protected by an in-memory
        TTL cache, bounded retries, and a circuit breaker.
      operationId: getBenefitByRef
      parameters:
        - name: ref
          in: path
          required: true
          description: Benefits Register reference, including its slashes.
          schema:
            type: string
          example: CA/2016/4001
      responses:
        '200':
          description: The benefit record.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BenefitRecord'
        '404':
          $ref: '#/components/responses/NotFound'
        '502':
          $ref: '#/components/responses/UpstreamUnavailable'

components:
  responses:
    NotFound:
      description: No matching record found in upstream source.
      content:
        text/plain:
          schema:
            type: string
          example: record not found
    UpstreamUnavailable:
      description: The upstream source failed or timed out.
      content:
        text/plain:
          schema:
            type: string
          example: service unavailable

  schemas:
    Resident:
      type: object
      properties:
        id:
          type: string
          example: R-10001
        first_name:
          type: string
          example: Ashley
        last_name:
          type: string
          example: Kessler
        date_of_birth:
          type: string
          example: '1983-01-23'
        address_line:
          type: string
          example: 203 Hazel St
        city:
          type: string
          example: Calder Central
        phone:
          type: string
          example: 555-102-3948
        program_status:
          type: string
          example: Active
        last_contact:
          type: string
          example: '2025-10-12'

    BenefitRecord:
      type: object
      properties:
        ref:
          type: string
          example: CA/2016/4001
        name:
          type: string
          example: KESSLER, Ashley
        born:
          type: string
          example: '1983-01-23'
        addr:
          type: string
          example: 203 Hazel Street
        town:
          type: string
          example: Calder Central
        benefit_code:
          type: string
          example: HSP-A
        review_due:
          type: string
          example: '2026-06-19'

    SourceStatus:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [ok, not_found, unavailable, circuit_open, stale]
          description: |
            ok — source answered successfully (or fresh cache hit);
            not_found — source answered, no such record;
            unavailable — source failed or timed out;
            circuit_open — circuit breaker tripped, failing fast;
            stale — serving expired cached record during outage.
          example: ok
        http_code:
          type: integer
          description: Last HTTP status code returned by upstream.
          example: 200
        latency_ms:
          type: integer
          format: int64
          description: Round-trip execution latency in milliseconds.
          example: 14
        error_message:
          type: string
          description: Diagnostic details when source status is not ok.
          example: upstream returned HTTP 500
        circuit:
          type: string
          enum: [closed, open, half_open]
          description: Circuit breaker state (Benefits Register only).
          example: closed

    HealthResponse:
      type: object
      properties:
        status:
          type: string
          enum: [ok, degraded]
          description: ok only when all sources are healthy.
          example: ok
        sources:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/SourceStatus'
          description: Live status per upstream source.

    UnifiedMeta:
      type: object
      properties:
        sources:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/SourceStatus'
          description: Status per queried source, keyed by source name.
        partial:
          type: boolean
          description: True when any queried source failed or returned degraded data.

    PaginationStatus:
      type: object
      properties:
        pages_fetched:
          type: integer
          example: 27
        records_seen:
          type: integer
          description: Total records received across all pages, duplicates included.
          example: 661
        duplicates:
          type: integer
          description: Number of duplicate records filtered out.
          example: 41
        conflicts:
          type: integer
          description: Duplicates whose payload differed from first-seen copy.
          example: 0
        unique:
          type: integer
          description: Distinct residents collected in first-seen order.
          example: 620
        reported_total:
          type: integer
          description: Total record count reported by the source.
          example: 620
        complete:
          type: boolean
          description: True when unique count equals reported total.
        reason:
          type: string
          description: Explains why the catalogue is incomplete, if applicable.

    ResidentListResponse:
      type: object
      properties:
        residents:
          type: array
          items:
            $ref: '#/components/schemas/Resident'
        pagination:
          $ref: '#/components/schemas/PaginationStatus'
        _meta:
          $ref: '#/components/schemas/UnifiedMeta'

    UnifiedResponse:
      type: object
      properties:
        resident:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/Resident'
          description: Present when resident_id was queried and found.
        benefits:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/BenefitRecord'
          description: Present when benefit_ref was queried and found.
        _meta:
          $ref: '#/components/schemas/UnifiedMeta'

    IdentityEvidence:
      type: object
      required: [rule]
      properties:
        rule:
          type: string
          enum: [exact_dob_and_name, name_town_street]
          description: Matching rule this evidence entry describes.
        resident_value:
          type: string
          description: Normalized resident attributes used for comparison.
          example: '1971-04-02 | delgado maria'
        benefit_value:
          type: string
          description: Normalized benefit-record attributes compared against.
          example: '1971-04-02 | delgado maria'
        result:
          type: string
          description: Outcome of the rule attempt when it did not merge.
          example: 0 candidates

    IdentityMatchMeta:
      type: object
      required: [outcome, candidate_refs, evidence, catalogue_fetched_at_ms]
      properties:
        outcome:
          type: string
          enum: [matched, ambiguous, no_match, unavailable]
          description: |
            matched — exactly one candidate cleared a rule;
            ambiguous — multiple candidates tied, declining to merge;
            no_match — searched the catalogue, nothing cleared any rule;
            unavailable — catalogue could not be obtained, no search happened.
        matched_ref:
          type: string
          description: Benefit reference of the merged record (matched only).
          example: CA/2016/4001
        candidate_refs:
          type: array
          items:
            type: string
          description: The winner (matched) or all tied candidates (ambiguous).
        evidence:
          type: array
          items:
            $ref: '#/components/schemas/IdentityEvidence'
          description: Audit trail of every rule attempted and what it saw.
        catalogue_fetched_at_ms:
          type: integer
          format: int64
          description: When the in-memory benefits catalogue was fetched (unix ms).

    AutoUnifiedResponse:
      type: object
      properties:
        resident:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/Resident'
        benefits:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/BenefitRecord'
          description: The linked benefit record; null unless outcome is matched.
        identity_match:
          $ref: '#/components/schemas/IdentityMatchMeta'
        _meta:
          $ref: '#/components/schemas/UnifiedMeta'